cast g_accuracy to float

+++ ppo/subtasks.py
@@ -144,7 +144,7 @@ class SubtasksAgent(Agent, NNBase):
         g_accuracy = torch.all(
             (hx.g - g_target[:, :, 0, 0]).abs() < 1e-3, dim=-1)
 
-        log = dict(g_accuracy=g_accuracy)
+        log = dict(g_accuracy=g_accuracy.float())
         for k, v in hx._asdict().items():
             if k.endswith('_loss'):
                 log[k] = v
